Added item position mod in animations Changed GuiButton Fixed GuiMain Revert hud font change Added a lame longjump

+++ src/minecraft/cat/util/MillisTimer.java
@@ -17,7 +17,7 @@ public final class MillisTimer {
         millis = System.currentTimeMillis();
     }
 
-    public boolean hasTicksPassed(final long ticks){
+    public boolean hasTicksPassed(final float ticks){
         return System.currentTimeMillis() >= millis + (ticks * 50);
     }
 }
